refactor: searchに用いられているhooksを一つに統合した
