+++ apps/web/src/components/TranscriptsSearchForm.tsx
@@ -1,23 +1,27 @@
 'use client';
 
-import { useTranscriptsSearch } from '~/hooks/useTranscriptsSearch';
 import type { Transcript } from '~/types/transcripts';
 import { Calendar } from '~/components/ui/calendar';
 import { Input } from '~/components/ui/input';
 import { Button } from '~/components/ui/button';
 
 type TranscriptsSearchFormProps = {
-  transcripts: Transcript[];
-  onResult: (results: Transcript[]) => void;
+  keyword: string;
+  setKeyword: (v: string) => void;
+  date: Date | undefined;
+  setDate: (d: Date | undefined) => void;
+  onSearch: () => void;
+  onReset: () => void;
 };
 
 export const TranscriptsSearchForm: React.FC<TranscriptsSearchFormProps> = ({
-  transcripts,
-  onResult,
+  keyword,
+  setKeyword,
+  date,
+  setDate,
+  onSearch,
+  onReset,
 }) => {
-  const { keyword, setKeyword, date, setDate, handleSearch } =
-    useTranscriptsSearch(transcripts, onResult);
-
   return (
     <div className="">
       <div className="flex flex-col justify-center items-center gap-4 mb-4">
@@ -28,12 +32,11 @@ export const TranscriptsSearchForm: React.FC<TranscriptsSearchFormProps> = ({
           onChange={(e) => setKeyword(e.target.value)}
         />
         <Calendar mode="single" selected={date} onSelect={setDate} />
-        <Button
-          type="button"
-          onClick={handleSearch}
-          className="w-60 h-10 text-lg"
-        >
+        <Button type="button" onClick={onSearch} className="w-60 h-10 text-lg">
           検索
+        </Button>
+        <Button type="button" onClick={onReset} className="w-60 h-10 text-lg">
+          リセット
         </Button>
       </div>
     </div>
